==> R/00_all.qmd <==
---
title: "00_all"
format: html
editor: visual
---

```{r}
#| label: setup
#| include: false

source("99_proj_func.R") # Import helper functions  
```

## Specified "/R" scripts

```{r}
#| label: Specified /R scripts
#| include: false

# list of files to run/render in /R
# Write the files in the list you want to render
# REMEMBER: add .qmd to the filename


qmd_files <- c(                 
  "01_load_data.qmd", 
  "02_clean_data.qmd",
  "07_correlations.qmd"
  )

#Run/render
render_qmd(qmd_files)
```

## ALL "/R" scripts

```{r}
#| label: ALL /R scripts
#| include: false

#list of files NOT to run (do not change).
not_run_qmd_files <- c(
  "00_all.qmd",
  "99_proj_func.R"
)

#files for rendering
all_qmd_files <- list.files(path = ".", full.names = FALSE) |> 
  setdiff(not_run_qmd_files)

render_qmd(all_qmd_files)
```
