==> R/01_load_data.qmd <==
---
title: "01_load_data"
format:
  html:
    embed-resources: true
knitr:
  opts_chunk:
    dev: "ragg_png"
    fig.path: "../results/"
editor: visual
---

```{r}
#| label: setup
#| message: false
#Load libraries
library("BiocGenerics")
library("Biobase")
library("GEOquery")
library(tidyverse)
```

# Create local directories

```{r}
# create local "_raw" and "data" directories if none exist.
dir.create(file.path("..", "_raw"),  recursive = TRUE, showWarnings = FALSE)
dir.create(file.path("..", "data"),  recursive = TRUE, showWarnings = FALSE)

```

# Download dataset

```{r}
#| label: Download dataset

#Download Data in ../_raw
input_url <- "https://ftp.ncbi.nlm.nih.gov/geo/datasets/GDS4nnn/GDS4337/soft/GDS4337_full.soft.gz"  #Download link
output_file <- "../_raw/full_dataset.soft.gz"                                                       #Output location + name
curl::curl_download(url = input_url, destfile = output_file)                                        
```

# Convert to csv and save in ../data

```{r}
#| label: convert & save to ../data

#Convert .soft file to table object and save in ../data
GEO_data <- getGEO(filename="../_raw/full_dataset.soft.gz") 
data_table <- Table(GEO_data)

```

# Save object table to /data

```{r}
write.csv(data_table, file="../data/01_data_table_load.csv")

#Extrack "disease.state" from .soft dataset
disease_state_vector <- Columns(GEO_data)$disease.state
Label_vector <- Columns(GEO_data)$sample

#convert to vector
disease_df <- tibble::tibble(
  disease.state = disease_state_vector,
  Sample = Label_vector
)
#Save as .csv in ../data
write.csv(disease_df, file="../data/01_disease_state_vector_load.csv")
```
