==> R/03_aug_data.qmd <==
---
title: "03_aug_data"
format:
  html:
    embed-resources: true
knitr:
  opts_chunk:
    dev: "ragg_png"
    fig.path: "../results/"
editor: visual
---

# Load libraries

```{r}
#| message: false
library("tidyverse")
```

# Augment data

## Load data

```{r}
clean_data <- read.csv("../data/02_clean_data.csv")
```

## Split data into two sub data sets

One data set containing only T2D subjects and one without T2D subjects

```{r}
# Filter disease identifier row
disease_row <- clean_data |> 
  filter(IDENTIFIER == "disease.state")

# Identify GSM-subjects with T2D
gsm_cols_T2D <- disease_row |> 
  select(starts_with("GSM")) |> 
  select(where(~ . == "T2D")) |> 
  names()

# Identify GSM-subjects without T2D
gsm_cols_nonT2D <- disease_row |> 
  select(starts_with("GSM")) |> 
  select(where(~ . == "non-diabetic")) |> 
  names()

# Remove the disease.state row 
df_full <- clean_data |> 
  filter(IDENTIFIER != "disease.state")

# Identify all other columns 
other_cols <- df_full |> 
  select(-starts_with("GSM")) |> 
  names()

# Combine the two data sets
df_T2D <- df_full |> 
  select(all_of(other_cols), all_of(gsm_cols_T2D))

df_non_T2D <- df_full |> 
  select(all_of(other_cols), all_of(gsm_cols_nonT2D))

# Make all GSM-values to numeric
df_T2D <- df_T2D |> 
  mutate(across(starts_with("GSM"), as.numeric))

df_non_T2D <- df_non_T2D |> 
  mutate(across(starts_with("GSM"), as.numeric))
```

Save the two data sets as csv-files

```{r}
write.csv(df_T2D, file="../data/03_T2D_data.csv")
write.csv(df_non_T2D, file="../data/03_none_T2D_data.csv")
```
